restructure test folder
