+++ test/models/event-tests.js
@@ -1,7 +1,7 @@
 import { assert } from "chai";
-import { db } from "../src/models/db.js";
-import { wellnessRetreat, testEvents, healthAndWellness } from "./fixtures.js";
-import { assertSubset } from "./test-utils.js";
+import { db } from "../../src/models/db.js";
+import { wellnessRetreat, testEvents, healthAndWellness } from "../fixtures.js";
+import { assertSubset } from "../test-utils.js";
 
 suite("Event Model tests", () => {
   let healthAndWellnessCat = null;
